separate token settings code
